Use non-animated splitter sprite for vertical wires
This simplifies the rendering code enormously for both renderers,
since all animated wires now move in a single direction.
It's also closer to the original Paradroid.

Actually, now that I think about it, this will finally enable us to
use a gradient pattern for active wires in the vector renderer.
So that's done too, which gets rid of that awful fake gradient
made from 5 dashed lines.